Rework history table to record intervals (#62)
This has been running on the Five Colleges instance for awhile and works well for the most part. Thinking about the future, the download, transform, and index portions might not all happen one after another. The thing that we really care about is the time it takes so that is what we're recording. The download time is still relevant as that is how fresh the data is. The start time is relevant to group different loads together.

This is a "breaking" change but only if someone has installed the unreleased v4 version.

The code around the database setup has also been refactored to keep the postgres error handling in postgres and slim down the _database/__init__.py file as it was hard to work in for this PR.
