+++ tests/test_load_history.py
@@ -1,6 +1,7 @@
 from collections.abc import Callable
 from contextlib import closing
 from dataclasses import dataclass
+from datetime import timedelta
 from typing import TYPE_CHECKING, cast
 from unittest import mock
 from unittest.mock import MagicMock
@@ -137,8 +138,14 @@ def _assert(
             assert (d := cur.fetchone()) is not None
             assert d[1] == "/patched"
             assert d[2] == q
-            assert d[7] == t
-            assert d[6] > d[5] > d[4] > d[3]
+            assert d[3] == t
+            assert d[4] > d[5]
+            assert d[6] > timedelta(microseconds=0)
+            assert d[6] < timedelta(seconds=1)
+            assert d[7] > timedelta(microseconds=0)
+            assert d[7] < timedelta(seconds=1)
+            assert d[8] > timedelta(microseconds=0)
+            assert d[8] < timedelta(seconds=1)
 
 
 @mock.patch("httpx_folio.auth.httpx.post")
